// gen/ts/open_splunk/dashboard_api.ts
export interface RunDashboardPanelRequest {
  dashboardId: string;
  panelId: string;
}

export interface RunDashboardPanelResponse {
  searchJob: SearchJob | undefined;
}
// ...
export const RunDashboardPanelRequest: MessageFns<RunDashboardPanelRequest> = {
  encode(message: RunDashboardPanelRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.dashboardId !== "") {
      writer.uint32(10).string(message.dashboardId);
    }
    if (message.panelId !== "") {
      writer.uint32(18).string(message.panelId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RunDashboardPanelRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const previousRecursionDepth = (reader as any).__tsProtoDecodeDepth ?? 0;
    if (previousRecursionDepth >= 100) {
      throw new globalThis.Error("protobuf decode recursion limit exceeded");
    }
    (reader as any).__tsProtoDecodeDepth = previousRecursionDepth + 1;
    try {
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseRunDashboardPanelRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.dashboardId = reader.string();
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.panelId = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    } finally {
      (reader as any).__tsProtoDecodeDepth = previousRecursionDepth;
    }
  },

  fromJSON(object: any): RunDashboardPanelRequest {
    return {
      dashboardId: isSet(object.dashboardId)
        ? globalThis.String(object.dashboardId)
        : isSet(object.dashboard_id)
        ? globalThis.String(object.dashboard_id)
        : "",
      panelId: isSet(object.panelId)
        ? globalThis.String(object.panelId)
        : isSet(object.panel_id)
        ? globalThis.String(object.panel_id)
        : "",
    };
  },

  toJSON(message: RunDashboardPanelRequest): unknown {
    const obj: any = {};
    if (message.dashboardId !== "") {
      obj.dashboardId = message.dashboardId;
    }
    if (message.panelId !== "") {
      obj.panelId = message.panelId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RunDashboardPanelRequest>, I>>(base?: I): RunDashboardPanelRequest {
    return RunDashboardPanelRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RunDashboardPanelRequest>, I>>(object: I): RunDashboardPanelRequest {
    const message = createBaseRunDashboardPanelRequest();
    message.dashboardId = object.dashboardId ?? "";
    message.panelId = object.panelId ?? "";
    return message;
  },
};

function createBaseRunDashboardPanelResponse(): RunDashboardPanelResponse {
  return { searchJob: undefined };
}

export const RunDashboardPanelResponse: MessageFns<RunDashboardPanelResponse> = {
  encode(message: RunDashboardPanelResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.searchJob !== undefined) {
      SearchJob.encode(message.searchJob, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RunDashboardPanelResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const previousRecursionDepth = (reader as any).__tsProtoDecodeDepth ?? 0;
    if (previousRecursionDepth >= 100) {
      throw new globalThis.Error("protobuf decode recursion limit exceeded");
    }
    (reader as any).__tsProtoDecodeDepth = previousRecursionDepth + 1;
    try {
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseRunDashboardPanelResponse();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.searchJob = SearchJob.decode(reader, reader.uint32());
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    } finally {
      (reader as any).__tsProtoDecodeDepth = previousRecursionDepth;
    }
  },

  fromJSON(object: any): RunDashboardPanelResponse {
    return {
      searchJob: isSet(object.searchJob)
        ? SearchJob.fromJSON(object.searchJob)
        : isSet(object.search_job)
        ? SearchJob.fromJSON(object.search_job)
        : undefined,
    };
  },

  toJSON(message: RunDashboardPanelResponse): unknown {
    const obj: any = {};
    if (message.searchJob !== undefined) {
      obj.searchJob = SearchJob.toJSON(message.searchJob);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RunDashboardPanelResponse>, I>>(base?: I): RunDashboardPanelResponse {
    return RunDashboardPanelResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RunDashboardPanelResponse>, I>>(object: I): RunDashboardPanelResponse {
    const message = createBaseRunDashboardPanelResponse();
    message.searchJob = (object.searchJob !== undefined && object.searchJob !== null)
      ? SearchJob.fromPartial(object.searchJob)
      : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string; value: unknown } ? { $case: T["$case"]; value?: DeepPartial<T["value"]> }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}